+++ src/capabilities/modules/module.h
@@ -15,7 +15,7 @@ public:
 
 public:
 
-    void animate(Uint8 mode, Uint8 def = Animator::DEFAULT_INACTIVE);
+    void animate(unsigned mode, unsigned def = Animator::DEFAULT_INACTIVE);
 
 
     // Worker interface
